fix: only lint and format staged changes
This commit adds the package lint-staged. I think when lint-staged
is calling pnpm run lint it passes some file arguments. Since pnpm
run lint calls turbo run lint this argument is getting passed directly
resulting in an error as extra arguments get added to the tubro command.
to prevent this we add '--' to signigy not to pass on further arguments.

:wq
:wq
:wq
wq
qw
:qw